RepairDesk.WpfClient/Converters/ImagePathConverter.cs
using System;
using System.Collections.Generic;
using System.Globalization;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using System.Windows;
using System.Windows.Data;
using Microsoft.Extensions.Configuration;


namespace RepairDesk.WpfClient.Converters
{
    public class ImagePathConverter : IValueConverter
    {
        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
        {
            var imageName = value as string;
            if (imageName != null)
            {
                var baseUrl = ((App)Application.Current).Configuration.GetValue<string>("WebConfig:PicturesUrlBase");
                return string.Format("{0}/{1}", baseUrl, imageName);
            }
            return null;
        }

        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
        {
            throw new NotImplementedException();
        }
    }
}
